refactor: use same alignment enum for percent and description of helper text
